Extract error handling to a concern
Group together error handling code in a specific concern to avoid growing too much the `ApplicationController`
