+++ app/controllers/application_controller.rb
@@ -1,5 +1,5 @@
 class ApplicationController < ActionController::Base
-  protect_from_forgery with: :exception
+  include ErrorHandling
 
   # :nocov:
   if ENV.fetch('HTTP_AUTH_ENABLED', false)
@@ -14,20 +14,6 @@ class ApplicationController < ActionController::Base
     payload[:referrer] = request&.referrer
     payload[:session_id] = request&.session&.id
     payload[:user_agent] = request&.user_agent
-  end
-
-  rescue_from Exception do |exception|
-    case exception
-    when Errors::InvalidSession, ActionController::InvalidAuthenticityToken
-      redirect_to invalid_session_errors_path
-    when Errors::ApplicationNotFound
-      redirect_to application_not_found_errors_path
-    else
-      raise if Rails.application.config.consider_all_requests_local
-
-      Raven.capture_exception(exception)
-      redirect_to unhandled_errors_path
-    end
   end
 
   helper_method :current_c100_application
@@ -46,9 +32,5 @@ class ApplicationController < ActionController::Base
     C100Application.create(attributes).tap do |c100_application|
       session[:c100_application_id] = c100_application.id
     end
-  end
-
-  def check_c100_application_presence
-    raise Errors::InvalidSession unless current_c100_application
   end
 end
